Refactoring the DesignTools primitives
