+++ DesignTools/Operations/SetCenter.cs
@@ -1,0 +1,74 @@
+﻿/*
+Copyright (c) 2018, Lars Brubaker, John Lewin
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+1. Redistributions of source code must retain the above copyright notice, this
+   list of conditions and the following disclaimer.
+2. Redistributions in binary form must reproduce the above copyright notice,
+   this list of conditions and the following disclaimer in the documentation
+   and/or other materials provided with the distribution.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+The views and conclusions contained in the software and documentation are those
+of the authors and should not be interpreted as representing official policies,
+either expressed or implied, of the FreeBSD Project.
+*/
+
+using MatterHackers.DataConverters3D;
+using MatterHackers.VectorMath;
+
+namespace MatterHackers.MatterControl.DesignTools.Operations
+{
+	public class SetCenter : Object3D
+	{
+		public SetCenter()
+		{
+		}
+
+		public SetCenter(IObject3D item, Vector3 position)
+		{
+			Matrix *= Matrix4X4.CreateTranslation(position - item.GetCenter());
+			Children.Add(item.Clone());
+		}
+
+		public SetCenter(IObject3D item, double x, double y, double z)
+			: this(item, new Vector3(x, y, z))
+		{
+		}
+
+		public SetCenter(IObject3D item, Vector3 offset, bool onX = true, bool onY = true, bool onZ = true)
+		{
+			var center = item.GetAxisAlignedBoundingBox(Matrix4X4.Identity).Center;
+
+			Vector3 consideredOffset = Vector3.Zero; // zero out anything we don't want
+			if (onX)
+			{
+				consideredOffset.X = offset.X - center.X;
+			}
+			if (onY)
+			{
+				consideredOffset.Y = offset.Y - center.Y;
+			}
+			if (onZ)
+			{
+				consideredOffset.Z = offset.Z - center.Z;
+			}
+
+			Matrix *= Matrix4X4.CreateTranslation(consideredOffset);
+			Children.Add(item.Clone());
+		}
+	}
+}
