+++ DesignTools/TestParts/BendOperation.cs
