+++ DesignTools/TestParts/RibonWithName.cs
@@ -1,0 +1,89 @@
+﻿/*
+Copyright (c) 2018, Lars Brubaker, John Lewin
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+1. Redistributions of source code must retain the above copyright notice, this
+   list of conditions and the following disclaimer.
+2. Redistributions in binary form must reproduce the above copyright notice,
+   this list of conditions and the following disclaimer in the documentation
+   and/or other materials provided with the distribution.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+The views and conclusions contained in the software and documentation are those
+of the authors and should not be interpreted as representing official policies,
+either expressed or implied, of the FreeBSD Project.
+*/
+
+using System;
+using System.ComponentModel;
+using System.Threading;
+using MatterHackers.Agg.Font;
+using MatterHackers.DataConverters3D;
+using MatterHackers.MatterControl.DesignTools.Operations;
+using MatterHackers.VectorMath;
+
+namespace MatterHackers.MatterControl.DesignTools
+{
+	public class RibonWithName : Object3D, IRebuildable
+	{
+		public override string ActiveEditor => "PublicPropertyEditor";
+
+		public RibonWithName()
+		{
+			Rebuild();
+		}
+
+		[DisplayName("Name")]
+		public string NameToWrite { get; set; } = "MatterHackers";
+
+		public NamedTypeFace Font { get; set; } = new NamedTypeFace();
+
+		public void Rebuild()
+		{
+			IObject3D cancerRibonStl = Object3D.Load("Cancer_Ribbon.stl", CancellationToken.None);
+
+			cancerRibonStl = new Rotate(cancerRibonStl, MathHelper.DegreesToRadians(90));
+
+			var letterPrinter = new TypeFacePrinter(NameToWrite.ToUpper(), new StyledTypeFace(NamedTypeFaceCache.GetTypeFace(Font), 12));
+
+			IObject3D nameMesh = new Object3D()
+			{
+				Mesh = VertexSourceToMesh.Extrude(letterPrinter, 5)
+			};
+
+			AxisAlignedBoundingBox textBounds = nameMesh.GetAxisAlignedBoundingBox();
+			var textArea = new Vector2(25, 6);
+
+			double scale = Math.Min(textArea.X / textBounds.XSize, textArea.Y / textBounds.YSize);
+			nameMesh = new Scale(nameMesh, scale, scale, 2 / textBounds.ZSize);
+			nameMesh = new Align(nameMesh, Face.Bottom | Face.Front, cancerRibonStl, Face.Top | Face.Front, 0, 0, -1);
+			nameMesh = new SetCenter(nameMesh, cancerRibonStl.GetCenter(), true, false, false);
+
+			nameMesh = new Rotate(nameMesh, 0, 0, MathHelper.DegreesToRadians(50));
+			nameMesh = new Translate(nameMesh, -37, -14, -1);
+
+			// output two meshes for card holder and text
+			this.Children.Modify(list =>
+			{
+				list.Clear();
+				list.Add(cancerRibonStl);
+				list.Add(nameMesh);
+			});
+
+			this.Mesh = null;
+		}
+	}
+}
