+++ MatterControl.csproj
@@ -86,9 +86,29 @@
     <Compile Include="ApplicationView\OemProfileDictionary.cs" />
     <Compile Include="ApplicationView\WindowsPlatformsFeatures.cs" />
     <Compile Include="DataStorage\ApplicationDataStorage.cs" />
-    <Compile Include="DesignTools\PackageOperation.cs" />
-    <Compile Include="DesignTools\MeshObjects.cs" />
+    <Compile Include="DesignTools\Operations\Object3DExtensions.cs" />
+    <Compile Include="DesignTools\Operations\Align.cs" />
+    <Compile Include="DesignTools\Operations\Scale.cs" />
+    <Compile Include="DesignTools\Operations\Rotate.cs" />
+    <Compile Include="DesignTools\Primitives\PyramidObject3D.cs" />
+    <Compile Include="DesignTools\Primitives\CylinderAdvancedObject3D.cs" />
+    <Compile Include="DesignTools\Primitives\CylinderObject3D.cs" />
+    <Compile Include="DesignTools\Primitives\SphereObject3D.cs" />
+    <Compile Include="DesignTools\Primitives\TextObject3D.cs" />
+    <Compile Include="DesignTools\Primitives\TorusObject3D.cs" />
+    <Compile Include="DesignTools\TestParts\TestPart.cs" />
+    <Compile Include="DesignTools\TestParts\RibonWithName.cs" />
+    <Compile Include="DesignTools\TestParts\PvcT.cs" />
+    <Compile Include="DesignTools\TestParts\PinchTest.cs" />
+    <Compile Include="DesignTools\TestParts\CurveTest.cs" />
+    <Compile Include="DesignTools\Primitives\ConeObject3D.cs" />
+    <Compile Include="DesignTools\Primitives\CubeObject3D.cs" />
+    <Compile Include="DesignTools\Operations\SetCenter.cs" />
+    <Compile Include="DesignTools\TestParts\BadSubtract.cs" />
+    <Compile Include="DesignTools\TestParts\CardHolder.cs" />
+    <Compile Include="DesignTools\Operations\Package.cs" />
     <Compile Include="DesignTools\PubicPropertyEditor.cs" />
+    <Compile Include="DesignTools\Operations\Translate.cs" />
     <Compile Include="PartPreviewWindow\AlignControls.cs" />
     <Compile Include="CustomWidgets\InlineTitleEdit.cs" />
     <Compile Include="PartPreviewWindow\LibraryBrowserPage.cs" />
@@ -339,7 +359,7 @@
     <Compile Include="SlicerConfiguration\Settings\ActiveSliceSettings.cs" />
     <Compile Include="SlicerConfiguration\Settings\PrinterSettings.cs" />
     <Compile Include="DataStorage\Classic\ClassicSqlitePrinterProfiles.cs" />
-    <Compile Include="DesignTools\BendOperation.cs" />
+    <Compile Include="DesignTools\TestParts\BendOperation.cs" />
     <Compile Include="Utilities\AuthenticationData.cs" />
     <Compile Include="Utilities\LimitCallFrequency.cs" />
     <Compile Include="Utilities\SelectedListItems.cs" />
